Compute module path relative to git repo root
The root project's pom.xml file now need not be located at the root of
the git repository. Invoke the plugin as follows for a non-root
pom.xml file.

mvn -f aggregators/all/pom.xml releaser:release

The modulesToRelease property must be provided relative to the root of
the git repository as before.